Refactor - adding middleware separately.
