+++ src/utils/middleware.js
@@ -1,0 +1,12 @@
+const logger = store => {
+    return next => {
+      return action => {
+        console.log('[Middleware] dispatching', action);
+        const result = next(action);
+        console.log('[Middleware] next state', store.getState())
+        return result;
+      }
+    }
+};
+
+export default logger;
